frontend: Move MetadataDisplay into its own component
This is so we abstract it more and later create a story for it.
