+++ frontend/src/components/common/Resource/Resource.tsx
@@ -1,8 +1,6 @@
 import chevronLeft from '@iconify/icons-mdi/chevron-left';
 import eyeIcon from '@iconify/icons-mdi/eye';
 import eyeOff from '@iconify/icons-mdi/eye-off';
-import menuDown from '@iconify/icons-mdi/menu-down';
-import menuUp from '@iconify/icons-mdi/menu-up';
 import { Icon } from '@iconify/react';
 import { Button, InputLabel } from '@material-ui/core';
 import Box from '@material-ui/core/Box';
@@ -10,25 +8,21 @@ import Divider from '@material-ui/core/Divider';
 import Grid, { GridProps } from '@material-ui/core/Grid';
 import IconButton from '@material-ui/core/IconButton';
 import Input, { InputProps } from '@material-ui/core/Input';
-import { makeStyles } from '@material-ui/core/styles';
 import { TextFieldProps } from '@material-ui/core/TextField';
-import Typography, { TypographyProps } from '@material-ui/core/Typography';
+import Typography from '@material-ui/core/Typography';
 import Editor from '@monaco-editor/react';
 import { Base64 } from 'js-base64';
 import _ from 'lodash';
 import * as monaco from 'monaco-editor';
 import React from 'react';
 import { Link as RouterLink, useLocation } from 'react-router-dom';
-import { ResourceClasses } from '../../../lib/k8s';
 import {
   KubeCondition,
   KubeContainer,
   KubeObject,
   KubeObjectInterface,
-  KubeOwnerReference,
 } from '../../../lib/k8s/cluster';
 import { createRouteURL, RouteURLProps } from '../../../lib/router';
-import { localeDate } from '../../../lib/util';
 import { useTypedSelector } from '../../../redux/reducers/reducers';
 import Loader from '../../common/Loader';
 import { SectionBox } from '../../common/SectionBox';
@@ -37,169 +31,9 @@ import SimpleTable, { NameValueTable, NameValueTableRow } from '../../common/Sim
 import Empty from '../EmptyContent';
 import { DateLabel, HoverInfoLabel, StatusLabel, StatusLabelProps } from '../Label';
 import Link, { LinkProps } from '../Link';
-import { LightTooltip } from '../Tooltip';
 import DeleteButton from './DeleteButton';
 import EditButton from './EditButton';
-
-const useStyles = makeStyles(theme => ({
-  metadataValueLabel: {
-    color: theme.palette.text.primary,
-    backgroundColor: theme.palette.metadataBgColor,
-    fontSize: '1.1em',
-    wordBreak: 'break-word',
-    paddingLeft: theme.spacing(1),
-    paddingRight: theme.spacing(1),
-  },
-}));
-
-interface MetadataDisplayProps {
-  resource: KubeObjectInterface;
-  extraRows?: NameValueTableRow[] | null;
-}
-
-export function MetadataDisplay(props: MetadataDisplayProps) {
-  const { resource, extraRows } = props;
-
-  function makeOwnerReferences(ownerReferences: KubeOwnerReference[]) {
-    if (!resource || ownerReferences === undefined) {
-      return undefined;
-    }
-
-    const numItems = ownerReferences.length;
-    if (numItems === 0) {
-      return undefined;
-    }
-
-    return ownerReferences.map((ownerRef, i) => (
-      <>
-        {ownerRef.kind in ResourceClasses ? (
-          <Link
-            routeName={new ResourceClasses[ownerRef.kind]().detailsRoute}
-            params={{ name: ownerRef.name, namespace: resource.metadata.namespace }}
-          >
-            {ownerRef.kind}: {ownerRef.name}
-          </Link>
-        ) : (
-          `${ownerRef.kind}: ${ownerRef.name}`
-        )}
-        {i < numItems - 1 && <br />}
-      </>
-    ));
-  }
-
-  const mainRows = ([
-    {
-      name: 'Name',
-      value: resource.metadata.name,
-    },
-    {
-      name: 'Namespace',
-      value: resource.metadata.namespace && resource.metadata.namespace,
-      hide: !resource.metadata.namespace,
-    },
-    {
-      name: 'Creation',
-      value: localeDate(resource.metadata.creationTimestamp),
-    },
-    {
-      name: 'Labels',
-      value: resource.metadata.labels && <MetadataDictGrid dict={resource.metadata.labels} />,
-      hide: !resource.metadata.labels,
-    },
-    {
-      name: 'Annotations',
-      value: resource.metadata.annotations && (
-        <MetadataDictGrid dict={resource.metadata.annotations} />
-      ),
-      hide: !resource.metadata.annotations,
-    },
-    {
-      name:
-        resource.metadata.ownerReferences && resource.metadata.ownerReferences.length > 1
-          ? 'Owner refs'
-          : 'Controlled by',
-      value: makeOwnerReferences(resource.metadata.ownerReferences || []),
-      hide: !resource.metadata.ownerReferences || resource.metadata.ownerReferences.length === 0,
-    },
-  ] as NameValueTableRow[]).concat(extraRows || []);
-
-  return <NameValueTable rows={mainRows} />;
-}
-
-interface MetadataDictGridProps {
-  dict: {
-    [index: string]: string;
-    [index: number]: string;
-  };
-  showKeys?: boolean;
-}
-
-export function MetadataDictGrid(props: MetadataDictGridProps) {
-  const classes = useStyles({});
-  const { dict, showKeys = true } = props;
-  const [expanded, setExpanded] = React.useState(false);
-  const defaultNumShown = 20;
-
-  const keys = Object.keys(dict || []);
-
-  const MetadataEntry = React.forwardRef((props: TypographyProps, ref: any) => {
-    return <Typography noWrap {...props} className={classes.metadataValueLabel} ref={ref} />;
-  });
-
-  function makeLabel(key: string | number) {
-    let fullText = dict[key];
-
-    if (showKeys) {
-      fullText = key + ': ' + fullText;
-    }
-
-    let shortText = fullText;
-
-    // Shorten the label manually because relying on the ellipsing methods
-    // was not working (it would correctly ellipse the text, but the width of it
-    // would still extend the area/section where the text is contained).
-    if (fullText.length > 50) {
-      shortText = fullText.substr(0, 50) + '…';
-    }
-
-    let labelComponent = <MetadataEntry>{shortText}</MetadataEntry>;
-
-    // If the full label is not being shown, use a tooltip to show the full text
-    // to the user (so they select it, etc.).
-    if (fullText.length !== shortText.length) {
-      labelComponent = <LightTooltip title={fullText} children={labelComponent} interactive />;
-    }
-    return labelComponent;
-  }
-
-  return (
-    <Grid container spacing={1} justify="flex-start">
-      {keys.length > defaultNumShown && (
-        <Grid item>
-          <IconButton onClick={() => setExpanded(!expanded)} size="small">
-            <Icon icon={expanded ? menuUp : menuDown} />
-          </IconButton>
-        </Grid>
-      )}
-      <Grid
-        container
-        item
-        justify="flex-start"
-        spacing={1}
-        style={{
-          maxWidth: '80%',
-        }}
-      >
-        {/* Limit the size to two entries until the user chooses to expand the whole section */}
-        {keys.slice(0, expanded ? keys.length : defaultNumShown).map((key, i) => (
-          <Grid key={i} item zeroMinWidth>
-            {makeLabel(key)}
-          </Grid>
-        ))}
-      </Grid>
-    </Grid>
-  );
-}
+import { MetadataDictGrid, MetadataDisplay } from './MetadataDisplay';
 
 interface ResourceLinkProps extends Omit<LinkProps, 'routeName' | 'params'> {
   name?: string;
