+++ frontend/src/components/common/Resource/index.tsx
@@ -1,1 +1,2 @@
+export * from './MetadataDisplay';
 export * from './Resource';
